Fix per #18
Sostituzione di tutte le instanze di double con unsigned long e tutte le instanze di unsigned int con unsigned long, modifica di alcuni nomi e rimozione consigli per l'input decimale

+++ Funzioni/ApriFile.hpp
@@ -1,7 +1,7 @@
-void ApriFile(record Array[LunghezzaDatabase], unsigned int &IndiceVuoto, ifstream &FileInput)
+void ApriFile(record Array[LunghezzaDatabase], unsigned long &IndiceVuoto, ifstream &FileInput)
 {
     //Dichiarazione variabili
-    unsigned int ContatoreRiga = 0, ContatoreDati = 0;
+    unsigned long ContatoreRiga = 0, ContatoreDati = 0;
     string RigaFile, Temp;
     stringstream RigaFileStream;
