+++ Funzioni/Ricerca.hpp
@@ -1,11 +1,11 @@
-void RicercaPrompt(record Array[LunghezzaDatabase], unsigned int IndiceVuoto)
+void RicercaPrompt(record Array[LunghezzaDatabase], unsigned long IndiceVuoto)
 {
-    unsigned int Scelta = 0;
+    unsigned long Scelta = 0;
     string Luogo;
     string Batt;
-    double Val;
+    unsigned long Val;
     int SceltaConcentrazione;
-    double Val1, Val2;
+    unsigned long Val1, Val2;
     string Tipo;
 
     cout<<"[1]Ricerca per Luogo..."<<endl;
@@ -14,7 +14,7 @@ void RicercaPrompt(record Array[LunghezzaDatabase], unsigned int IndiceVuoto)
     cout<<"[4]Ricerca per Tipologia..."<<endl;
     cout<<endl;
     cout<<"Inserisci che tipo di ricerca vuoi fare: ";
-    InputInt(Scelta);
+    InputLong(Scelta);
     switch (Scelta)
     {
         case 1: system("cls");
@@ -41,21 +41,21 @@ void RicercaPrompt(record Array[LunghezzaDatabase], unsigned int IndiceVuoto)
                 case 1: system("cls");
                         cout<<"Hai selezionato la ricerca per concentrazione minore di un valore inserito"<<endl;
                         cout<<"Inserisci il valore: ";
-                        InputDouble(Val);
+                        InputLong(Val);
                         ConcentrazioneMinore(Array, IndiceVuoto, Val);
                 break;
                 case 2: system("cls");
                         cout<<"Hai selezionato la ricerca per concentrazione maggiore di un valore inserito"<<endl;
                         cout<<"Inserisci il valore: ";
-                        InputDouble(Val);
+                        InputLong(Val);
                         ConcentrazioneMaggiore(Array, IndiceVuoto, Val);
                 break;
                 case 3: system("cls");
                         cout<<"Hai selezionato la ricerca per concentrazione compresa tra due valori"<<endl;
                         cout<<"Inserisci il valore 1: ";
-                        InputDouble(Val1);
+                        InputLong(Val1);
                         cout<<"Inserisci il valore 2: ";
-                        InputDouble(Val2);
+                        InputLong(Val2);
                         if(Val1<Val2){
                                 ConcentrazioneCompresa(Array, IndiceVuoto, Val1, Val2);
                         }
@@ -79,7 +79,7 @@ void RicercaPrompt(record Array[LunghezzaDatabase], unsigned int IndiceVuoto)
 
 }
 
-void RicercaLuogo(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, string Localita)
+void RicercaLuogo(record Array[LunghezzaDatabase], unsigned long IndiceVuoto, string Localita)
 {
     int Conta=0;
     for(int i=0; i<LunghezzaDatabase; i++){
@@ -96,7 +96,7 @@ void RicercaLuogo(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, str
     }
 }
 
-void RicercaBatterio(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, string BatterioRicerca)
+void RicercaBatterio(record Array[LunghezzaDatabase], unsigned long IndiceVuoto, string BatterioRicerca)
 {
     int Conta=0;
     bool Ris=false;
@@ -118,7 +118,7 @@ void RicercaBatterio(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, 
     }
 }
 
-void ConcentrazioneMinore(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, double Valore)
+void ConcentrazioneMinore(record Array[LunghezzaDatabase], unsigned long IndiceVuoto, unsigned long Valore)
 {
     bool Ris=false;
     int Conta=0;
@@ -140,7 +140,7 @@ void ConcentrazioneMinore(record Array[LunghezzaDatabase], unsigned int IndiceVu
     }
 }
 
-void ConcentrazioneMaggiore(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, double Valore)
+void ConcentrazioneMaggiore(record Array[LunghezzaDatabase], unsigned long IndiceVuoto, unsigned long Valore)
 {
     int Conta=0;
     bool Ris=false;
@@ -163,7 +163,7 @@ void ConcentrazioneMaggiore(record Array[LunghezzaDatabase], unsigned int Indice
     }
 }
 
-void ConcentrazioneCompresa(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, double Valore1, double Valore2)
+void ConcentrazioneCompresa(record Array[LunghezzaDatabase], unsigned long IndiceVuoto, unsigned long Valore1, unsigned long Valore2)
 {
     int Conta=0;
     bool Ris=false;
@@ -185,7 +185,7 @@ void ConcentrazioneCompresa(record Array[LunghezzaDatabase], unsigned int Indice
     }
 }
 
-void RicercaTipo(record Array[LunghezzaDatabase], unsigned int IndiceVuoto, string Tipo)
+void RicercaTipo(record Array[LunghezzaDatabase], unsigned long IndiceVuoto, string Tipo)
 {
     int Conta=0;
     bool Ris=false;
